requirements.txt
Appium-Python-Client==2.6.1
PyHamcrest==2.0.4
pytest==7.1.2
allure-pytest==2.10.0